+++ Assets/Level Selection/testScript.cs
@@ -1,0 +1,9 @@
+using UnityEngine;
+
+public class TestButton : MonoBehaviour
+{
+    public void OnButtonPressed()
+    {
+        Debug.Log("Buton apasat!");
+    }
+}
