+++ Assets/Level Selection/testScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e92b90e18ae124c1d8318047c15067e0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
