+++ Assets/Scenes/Level 2.unity.meta
@@ -1,6 +1,5 @@
 fileFormatVersion: 2
-guid: 07b12cb605c984c909bb967ba035a0de
-folderAsset: yes
+guid: 7d3e48ed3a30c4e578f995aa40300363
 DefaultImporter:
   externalObjects: {}
   userData:
