+++ Assets/Sprites/Backgrounds.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0a6597a03d9b54ae5b96c91497e91dd1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
